src/test/java/com/unrc/app/PostTest.java
package com.unrc.app;

import com.unrc.app.models.Post;

import org.javalite.activejdbc.Base;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import static org.javalite.test.jspec.JSpec.the;
import static org.junit.Assert.assertEquals;

public class PostTest {
    @Before
    public void before(){
        Base.open("com.mysql.jdbc.Driver", "jdbc:mysql://localhost/carsapp_test", "root", "");
        System.out.println("PostTest setup");
        Base.openTransaction();
    }

    @After
    public void after(){
        System.out.println("PostTest tearDown");
        Base.rollbackTransaction();
        Base.close();
    }

    @Test
    public void shouldValidateMandatoryFields(){
        Post post = new Post();
        the(post).shouldNotBe("valid");

        the(post.errors().get("price")).shouldBeEqual("value is missing");

        post.set("price","25000");

        the(post).shouldBe("valid");
    } 
}